Add files via upload
 House Price Prediction with 12 predictor variable and one dependent variable which is the "Price."
Recursive feature elimination (RFE) is a feature selection method that fits a model and removes the weakest feature (or features) until the specified number of features is reached. ... RFE requires a specified number of features to keep, however it is often not known in advance how many features are valid.